infra: retryable get_mender_gateway
gateway is a bit special - right after 'up' it may still
be unavailable for some time (although nothing is wrong).

we can easily fail to get the ip at a first shot - so
do retry the query a couple times.

this together with wait_for_traefik should be enough
to always find the gateway and have a consistent setup.

changelog: none

Signed-off-by: Marcin Chalczynski <m.chalczynski@gmail.com>